Role and Permissions endpoint and models completed

+++ Vextech API/Models/PermissionModel.cs
@@ -2,7 +2,7 @@
 {
     public class PermissionModel
     {
-        public int ID { get; set; }
+        public ulong ID { get; set; }
         public string Name { get; set; }
     }
 }
